+++ src/ops/splice.rs
@@ -117,35 +117,3 @@ where
         }
     }
 }
-
-
-#[allow(suspicious_auto_trait_impls)]
-unsafe impl<'a, Traits: ?Sized + Send + Trait, M: MemBuilder, ReplaceIter: ExactSizeIterator> Send
-for
-    Splice<'a, AnyVecPtr<Traits, M>, ReplaceIter>
-where
-    ReplaceIter::Item: AnyValue + Send
-{}
-#[allow(suspicious_auto_trait_impls)]
-unsafe impl<'a, Type: Send, M: MemBuilder, ReplaceIter: ExactSizeIterator> Send
-for
-    Splice<'a, AnyVecRawPtr<Type, M>, ReplaceIter>
-where
-    ReplaceIter::Item: AnyValue + Send
-{}
-
-
-#[allow(suspicious_auto_trait_impls)]
-unsafe impl<'a, Traits: ?Sized + Sync + Trait, M: MemBuilder, ReplaceIter: ExactSizeIterator> Sync
-for
-    Splice<'a, AnyVecPtr<Traits, M>, ReplaceIter>
-where
-    ReplaceIter::Item: AnyValue + Sync
-{}
-#[allow(suspicious_auto_trait_impls)]
-unsafe impl<'a, Type: Sync, M: MemBuilder, ReplaceIter: ExactSizeIterator> Sync
-for
-    Splice<'a, AnyVecRawPtr<Type, M>, ReplaceIter>
-where
-    ReplaceIter::Item: AnyValue + Sync
-{}